Write runtimes of .apsimx files to configuration and use these to order Jenkins runs
