+++ UserInterface/Presenters/ExplorerPresenter.cs
@@ -220,9 +220,9 @@ namespace UserInterface.Presenters
                 try
                 {
                     if (this.ApsimXFile.FileName != null)
-                        Utility.Configuration.Settings.DelMruFile(this.ApsimXFile.FileName);
+                        Configuration.Settings.DelMruFile(this.ApsimXFile.FileName);
 
-                    Utility.Configuration.Settings.AddMruFile(newFileName);
+                    Configuration.Settings.AddMruFile(newFileName);
                     MainPresenter.ChangeTabText(this.view, Path.GetFileNameWithoutExtension(newFileName), newFileName);
                     this.ApsimXFile.Write(newFileName);
                     return true;
